pyproject.toml
[project]
name = "cerulean"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Anish Pallati", email = "anishp@duck.com" }]
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.6",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.1",
    "uvicorn>=0.32.1",
    "beautifulsoup4>=4.12.3",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.8.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cerulean"]
